Move sub-classes of ToolStripItem class to separate file
